requirements.txt
streamlit
transformers
sentence-transformers
scikit-learn
spotipy
python-dotenv
matplotlib
pandas
numpy
